Intermediate+_days_45_to_57/47_exercises/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
lxml==4.6.2
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
PyJWT==2.0.0
pyperclip==1.8.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.1
requests-cache==0.5.2
six==1.15.0
soupsieve==2.1
spotipy==2.16.1
twilio==6.50.1
urllib3==1.26.2
